reciever/requirements.txt
kafka-python
opencv-python
confluent-kafka
Pillow
protobuf
